% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_nj_tree}
\alias{build_nj_tree}
\title{Build a Neighbor-Joining tree from a multiple sequence alignment}
\usage{
build_nj_tree(msa, model = "raw", pairwise.deletion = TRUE)
}
\arguments{
\item{msa}{A multiple alignment object (class `MsaDNAMultipleAlignment` or similar)}

\item{model}{Evolutionary model for distance calculation passed to `ape::dist.dna`
(e.g., "raw", "JC69", "K80", etc.)}

\item{pairwise.deletion}{Logical. If TRUE, compute distances with pairwise deletion}
}
\value{
An object of class `phylo` (NJ tree)
}
\description{
This function takes a Multiple Sequence Alignment (MSA) object (e.g., output of
`align_sequences(method = "msa")`) and generates a Neighbor-Joining (NJ) tree.
}
