% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastr.R
\name{fetch_metadata}
\alias{fetch_metadata}
\title{Fetch Metadata (and optionally sequence ranges) from NCBI}
\usage{
fetch_metadata(accessions, db = c("nuccore", "protein"), seq_range = NULL)
}
\arguments{
\item{accessions}{Character vector of accession numbers.}

\item{db}{Either "nuccore" or "protein".}

\item{seq_range}{Either:
- NULL (default): fetch full sequence for every accession
- numeric(2):   fetch that same start–end for *all* accessions
- named list:   each element is a numeric(2) vector, names are accessions;
                will fetch only that slice for the named accession,
                full sequence for others.}
}
\value{
A tibble with columns
  `accession`, `accession_version`, `title`, `organism`, `sequence`
}
\description{
Fetch Metadata (and optionally sequence ranges) from NCBI
}
